Add Forum Pages - Add Forum Thanks Feature
Had to add all the changes to files to allow Forum Pages to be the next distinct feature page set. This included a bit of tweak to how check.ts works with pages and sub pages.

Then added the feature to forum.ts to gift FL wedges on forum posts as thank you with custom message.